requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-google-genai
langchain-community
langchain-huggingface
sentence-transformers
faiss-cpu
pypdf
python-dotenv